+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - BulkIngestRequest contains now one single list "items"
 - tests for ingestion adapted to BulkIngestRequest-Model
+- remove stop-gaps for MX-1596
+
 ### Deprecated
 
 ### Removed